+++ tiledb/sm/filesystem/posix.cc
@@ -166,7 +166,7 @@ Status Posix::create_dir(const std::string& path) const {
   }
 
   uint32_t permissions = 0;
-  RETURN_NOT_OK(get_posix_permissions(&permissions));
+  RETURN_NOT_OK(get_posix_directory_permissions(&permissions));
 
   if (mkdir(path.c_str(), permissions) != 0) {
     return LOG_STATUS(Status::IOError(
@@ -178,7 +178,7 @@ Status Posix::create_dir(const std::string& path) const {
 
 Status Posix::touch(const std::string& filename) const {
   uint32_t permissions = 0;
-  RETURN_NOT_OK(get_posix_permissions(&permissions));
+  RETURN_NOT_OK(get_posix_file_permissions(&permissions));
 
   int fd = ::open(filename.c_str(), O_WRONLY | O_CREAT | O_SYNC, permissions);
   if (fd == -1 || ::close(fd) != 0) {
@@ -435,15 +435,16 @@ Status Posix::read(
 
 Status Posix::sync(const std::string& path) {
   uint32_t permissions = 0;
-  RETURN_NOT_OK(get_posix_permissions(&permissions));
 
   // Open file
   int fd = -1;
-  if (is_dir(path))  // DIRECTORY
+  if (is_dir(path)) {  // DIRECTORY
+    RETURN_NOT_OK(get_posix_directory_permissions(&permissions));
     fd = open(path.c_str(), O_RDONLY, permissions);
-  else if (is_file(path))  // FILE
+  } else if (is_file(path)) {  // FILE
+    RETURN_NOT_OK(get_posix_file_permissions(&permissions));
     fd = open(path.c_str(), O_WRONLY | O_APPEND | O_CREAT, permissions);
-  else
+  } else
     return Status::Ok();  // If file does not exist, exit
 
   // Handle error
@@ -484,7 +485,7 @@ Status Posix::write(
   assert(found);
 
   uint32_t permissions = 0;
-  RETURN_NOT_OK(get_posix_permissions(&permissions));
+  RETURN_NOT_OK(get_posix_file_permissions(&permissions));
 
   // Get file offset (equal to file size)
   Status st;
@@ -579,11 +580,24 @@ Status Posix::write_at(
   return Status::Ok();
 }
 
-Status Posix::get_posix_permissions(uint32_t* permissions) const {
+Status Posix::get_posix_file_permissions(uint32_t* permissions) const {
   // Get config params
   bool found = false;
   std::string posix_permissions =
-      config_.get().get("vfs.file.posix_permissions", &found);
+      config_.get().get("vfs.file.posix_file_permissions", &found);
+  assert(found);
+
+  // Permissions are passed in octal notation by the user
+  *permissions = std::strtol(posix_permissions.c_str(), NULL, 8);
+
+  return Status::Ok();
+}
+
+Status Posix::get_posix_directory_permissions(uint32_t* permissions) const {
+  // Get config params
+  bool found = false;
+  std::string posix_permissions =
+      config_.get().get("vfs.file.posix_directory_permissions", &found);
   assert(found);
 
   // Permissions are passed in octal notation by the user